Move TimeCodeWrapper to own file.
