+++ MainWindow.xaml.cs
@@ -8,41 +8,6 @@ using Path = System.IO.Path;
 
 namespace VideoTrimmer
 {
-    public class TimeCodeWrapper
-    {
-        public bool IsStartMarker { get; private set; }
-        public System.Windows.Controls.TextBox TextBox { get; private set; }
-        public System.Windows.Controls.Button ButtonSet { get; private set; }
-        public System.Windows.Controls.Button ButtonJumpTo { get; private set; }
-        public TimeSpan? CurrentTime { get; private set; }
-
-        public TimeCodeWrapper(System.Windows.Controls.TextBox textBox, System.Windows.Controls.Button buttonSet, System.Windows.Controls.Button buttonJumpTo, bool isStartMarker)
-        {
-            TextBox = textBox;
-            ButtonSet = buttonSet;
-            ButtonJumpTo = buttonJumpTo;
-            IsStartMarker = isStartMarker;
-
-            SetTimeSpan(null);
-        }
-
-        public void SetTimeSpan(TimeSpan? timeSpan)
-        {
-            CurrentTime = timeSpan;
-            TextBox.Text = CurrentTime.HasValue ? CurrentTime.Value.ToString(Globals.timeFormat) : "N/A";
-            ButtonSet.IsEnabled = CurrentTime.HasValue;
-            ButtonJumpTo.IsEnabled = CurrentTime.HasValue;
-        }
-
-        public void SetLockStatus(bool newLockStatus)
-        {
-            TextBox.IsEnabled = newLockStatus;
-            ButtonSet.IsEnabled = newLockStatus;
-            ButtonJumpTo.IsEnabled = !newLockStatus;
-            CurrentTime = null;
-        }
-    }
-
     /// <summary>
     /// Interaction logic for MainWindow.xaml
     /// </summary>
