+++ VideoTrimmer.csproj
@@ -81,6 +81,7 @@
     <Compile Include="ProgressWindow.xaml.cs">
       <DependentUpon>ProgressWindow.xaml</DependentUpon>
     </Compile>
+    <Compile Include="TimeCodeWrapper.cs" />
     <Compile Include="VideoProcessing.cs" />
     <Page Include="MainWindow.xaml">
       <Generator>MSBuild:Compile</Generator>
